refactor/organize controllers
