+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/converter/interviewer/InterviewerAvailabilityToAvailabilityDto.java
@@ -1,9 +1,7 @@
 package com.github.TheDreigon.JavaInterviewCalendarAPI.converter.interviewer;
 
 import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.AvailabilityDto;
-import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.interviewer.InterviewerAvailabilityDto;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model.InterviewerAvailability;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
