+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/converter/interviewer/InterviewerToInterviewerDto.java
@@ -1,7 +1,6 @@
 package com.github.TheDreigon.JavaInterviewCalendarAPI.converter.interviewer;
 
 import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.AvailabilityDto;
-import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.interviewer.InterviewerAvailabilityDto;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.interviewer.InterviewerDto;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model.Interviewer;
 import org.springframework.beans.factory.annotation.Autowired;
