+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/service/impl/CandidateAvailabilityServiceImpl.java
@@ -1,7 +1,7 @@
 package com.github.TheDreigon.JavaInterviewCalendarAPI.service.impl;
 
-import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.candidate.CandidateAvailabilityDto;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.converter.candidate.CandidateAvailabilityToCandidateAvailabilityDto;
+import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.candidate.CandidateAvailabilityDto;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.exception.AvailabilityNotFoundException;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.exception.CandidateNotFoundException;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model.CandidateAvailability;
