+++ src/main/java/com/github/TheDreigon/JavaInterviewCalendarAPI/service/impl/InterviewerAvailabilityServiceImpl.java
@@ -1,8 +1,8 @@
 package com.github.TheDreigon.JavaInterviewCalendarAPI.service.impl;
 
-import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.interviewer.InterviewerAvailabilityDto;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.converter.interviewer.InterviewerAvailabilityDtoToInterviewerAvailability;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.converter.interviewer.InterviewerAvailabilityToInterviewerAvailabilityDto;
+import com.github.TheDreigon.JavaInterviewCalendarAPI.dto.interviewer.InterviewerAvailabilityDto;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.exception.AvailabilityNotFoundException;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.exception.InterviewerNotFoundException;
 import com.github.TheDreigon.JavaInterviewCalendarAPI.persistence.model.InterviewerAvailability;
